Added enclosed_roots_arg
Moved the old method of calculating the number of roots within a
contour based on the difference of the continued argument of f(z) to a
separate function for future experimentation.  The method of
approximating f’(z) with a Taylor series seems like a much more
reliable method in the case where f’(z) is not user defined since it
converges smoothly.