Approximate f'(z) for the Cauchy integral by a finite difference formula
It is not clear how to reliably determine if and when the Taylor
expansion for f’(z) converges, particularly if the Taylor coefficients
are computed by contour integration around a very large, contour.  This
method of approximating derivatives has been replaced with a finite
difference approximation which, although more prone to round-off
errors, is more certain to converge, even for very large contours.