Redo subdivisions when roots close to box
It may turn out after computing the roots that a root is very close to the contour which is supposed to contain it.  This makes the integration suspect so the multiplicity is computed directly by integrating in a small contour around the root.  The box containing the root and any sibling boxes created during the same subdivision are discarded.  The parent box is then returned to the list of boxes to be subdivided again.  The box which was too close to a root is added to a list of failedBoxes which the subdivision routine will skip.